fix vertex structures & add scene to model loader
