+++ include/render/render.h
@@ -30,7 +30,7 @@
 
 #include <shadermanager/ShaderManager.h>
 #include <modelloader/modelloader.h>
-#include <modelloader/Vertex.h>
+
 
 const uint32_t WIDTH = 800;
 const uint32_t HEIGHT = 600;
@@ -39,7 +39,6 @@ const int MAX_FRAMES_IN_FLIGHT = 3;
 const std::string MODEL_PATH = "misc/cube.obj";
 const std::string TEXTURE_PATH = "misc/white.jpg";
 const std::string MTL_PATH = "misc/";
-nevk::Model testmodel;
 
 const std::vector<const char*> validationLayers = {
     "VK_LAYER_KHRONOS_validation"
@@ -101,6 +100,86 @@ struct UniformBufferObject
     //alignas(16) glm::mat4 view;
     //alignas(16) glm::mat4 proj;
 };
+
+struct Vertex
+{
+    glm::vec3 pos;
+    glm::vec3 color;
+    glm::vec3 ka;
+    glm::vec3 kd;
+    glm::vec3 ks;
+    glm::vec2 texCoord;
+
+    static VkVertexInputBindingDescription getBindingDescription()
+    {
+        VkVertexInputBindingDescription bindingDescription{};
+        bindingDescription.binding = 0;
+        bindingDescription.stride = sizeof(Vertex);
+        bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
+
+        return bindingDescription;
+    }
+
+    static std::array<VkVertexInputAttributeDescription, 5> getAttributeDescriptions()
+    {
+        std::array<VkVertexInputAttributeDescription, 5> attributeDescriptions{};
+
+        attributeDescriptions[0].binding = 0;
+        attributeDescriptions[0].location = 0;
+        attributeDescriptions[0].format = VK_FORMAT_R32G32B32_SFLOAT;
+        attributeDescriptions[0].offset = offsetof(Vertex, pos);
+
+        attributeDescriptions[1].binding = 0;
+        attributeDescriptions[1].location = 1;
+        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
+        attributeDescriptions[1].offset = offsetof(Vertex, ka);
+
+        attributeDescriptions[2].binding = 0;
+        attributeDescriptions[2].location = 2;
+        attributeDescriptions[2].format = VK_FORMAT_R32G32B32_SFLOAT;
+        attributeDescriptions[2].offset = offsetof(Vertex, kd);
+
+        attributeDescriptions[3].binding = 0;
+        attributeDescriptions[3].location = 3;
+        attributeDescriptions[3].format = VK_FORMAT_R32G32B32_SFLOAT;
+        attributeDescriptions[3].offset = offsetof(Vertex, ks);
+
+        attributeDescriptions[4].binding = 0;
+        attributeDescriptions[4].location = 4;
+        attributeDescriptions[4].format = VK_FORMAT_R32G32_SFLOAT;
+        attributeDescriptions[4].offset = offsetof(Vertex, texCoord);
+
+        return attributeDescriptions;
+    }
+
+    bool operator==(const Vertex& other) const
+    {
+        return pos == other.pos && ka == other.ka && kd == other.kd && ks == other.ks && texCoord == other.texCoord;
+    }
+
+    static std::vector<Vertex> create(std::vector<nevk::Vertex> const& params)
+    {
+        std::vector<Vertex> ret(params.size());
+        std::transform(params.begin(), params.end(), ret.begin(),
+                       [](auto value) {
+                           return Vertex{value.pos, value.color, value.ka, value.kd, value.ks, value.uv};
+                       });
+        return ret;
+    }
+};
+
+namespace std
+{
+template <>
+struct hash<Vertex>
+{
+    size_t operator()(Vertex const& vertex) const
+    {
+        return ((hash<glm::vec3>()(vertex.pos) ^ (hash<glm::vec3>()(vertex.color) << 1)) >> 1) ^ (hash<glm::vec2>()(vertex.texCoord) << 1);
+    }
+};
+} // namespace std
+
 
 class Render
 {
@@ -173,6 +252,7 @@ private:
     bool framebufferResized = false;
 
     nevk::ShaderManager mShaderManager;
+    nevk::Scene mScene;
 
     void initWindow()
     {
@@ -1081,8 +1161,9 @@ private:
 
     void loadModels()
     {
-        testmodel.loadModel(MODEL_PATH, MTL_PATH);
-        vertices = testmodel.getVertices();
+        nevk::Model testmodel;
+        testmodel.loadModel(MODEL_PATH, MTL_PATH, mScene);
+        vertices = Vertex::create(testmodel.getVertices());
         indices = testmodel.getIndices();
     }
